+++ ApertureMapModelTools/__init__.py
@@ -15,7 +15,7 @@ from .__core__ import DataField, FractureImageStack
 from .__core__ import _get_logger, set_main_logger_level
 from .__core__ import files_from_directory, load_infile_list
 from .__core__ import calc_percentile, calc_percentile_num, get_data_vect
-from . import DataProcessing
+from . import data_processing
 from . import RunModel
 from . import OpenFoam
