+++ test/unit/TestDataProcessing/TestPercentiles.py
@@ -8,7 +8,7 @@ Last Modifed: 2016/10/25
 """
 import argparse
 import os
-from ApertureMapModelTools.DataProcessing.percentiles import Percentiles
+from ApertureMapModelTools.data_processing.percentiles import Percentiles
 
 
 class TestPercentiles:
